requirements.txt
Django==2.0
djangorestframework==3.11.0
pytz==2020.1